requirements.txt
torch
torchvision
transformers
tqdm
pillow
numpy
scikit-learn
matplotlib
